Move components to folder within lib
