+++ lib/components/Footer.js
