+++ lib/components/Heading.js
