+++ lib/components/NavBar.js
