+++ lib/components/OpenTargetsTitle.js
