+++ lib/components/Search.js
