+++ lib/components/SearchExamples.js
