+++ lib/components/Splash.js
