+++ lib/components/TestComponent.js
