+++ lib/index.js
@@ -11,18 +11,18 @@
 //     Row,
 // };
 
-import 'typeface-roboto';
+import "typeface-roboto";
 
-export { default as TestComponent } from './TestComponent';
-export { default as VirtualizedTable } from './VirtualizedTable';
-export { default as VirtualizedColumn } from './VirtualizedColumn';
-export { default as NavBar } from './NavBar';
-export { default as Footer } from './Footer';
-export { default as Page } from './Page';
-export { default as Card } from './Card';
-export { default as Heading } from './Heading';
-export { default as Splash } from './Splash';
-export { default as HomeBox } from './HomeBox';
-export { default as Search } from './Search';
-export { default as SearchExamples } from './SearchExamples';
-export { default as OtUiThemeProvider } from './OtUiThemeProvider';
+export { default as TestComponent } from "./components/TestComponent";
+export { default as VirtualizedTable } from "./components/VirtualizedTable";
+export { default as VirtualizedColumn } from "./components/VirtualizedColumn";
+export { default as NavBar } from "./components/NavBar";
+export { default as Footer } from "./components/Footer";
+export { default as Page } from "./components/Page";
+export { default as Card } from "./components/Card";
+export { default as Heading } from "./components/Heading";
+export { default as Splash } from "./components/Splash";
+export { default as HomeBox } from "./components/HomeBox";
+export { default as Search } from "./components/Search";
+export { default as SearchExamples } from "./components/SearchExamples";
+export { default as OtUiThemeProvider } from "./components/OtUiThemeProvider";
